refactor(lighting): compute per-instance light lists on the GPU
Replace the O(instances x lights) CPU culling in BLASCluster with a GPU compute pass. Each instance counts overlapping lights, atomically reserves a contiguous range in a shared index list, and writes the light indices, removing the 32-light-per-instance cap.

InstanceLightData is now a 16-bit offset/count pair into the GPU list, shrinking InstanceData from 144 to 112 bytes. Clusters upload only a per-instance world bound. LIGHTS_MAX is raised to 1024 and Light::m_Index widened to uint16_t (also fixes a latent uint8_t overflow in LightTLAS).

InstanceLightList is bound as an SRV at t12 (PT/SHaRC) and t17 (GI/SHaRCGI); the GLOBAL_LIGHTS path is preserved.

+++ src/Core/Light.h
@@ -11,7 +11,7 @@ struct Light
 	bool m_Active;
 
 	// Light data index
-	uint8_t m_Index;
+	uint16_t m_Index;
 
 	// Instances that are affected by the light, calculated from instance node worldBound center + radius and light position + radius
 	eastl::hash_set<BLASCluster*> m_Instances;
